Defined optional workaround for when uname is MINGW64_NT-10.0-19042

+++ jms-publisher/test-producer-jms-to-kakfa.sh
@@ -2,10 +2,14 @@
 #
 #java -classpath ~/.m2/repository/javax/jms/jms/1.1/jms-1.1.jar:kafka-jms-fatjar/target/kafka-jms-client-fat-4.1.2.jar:target/jmstokafka-0.0.1.jar com.fuelbigdata.publisher.ProducerApp
 #
-mkdir -p plibjava
-cp -rf  ~/.m2/repository/javax/jms/jms/1.1/jms-1.1.jar plibjava/
-cp -rf kafka-jms-fatjar/target/kafka-jms-client-fat-4.1.2.jar plibjava/
-cp -rf target/jmstokafka-0.0.1.jar plibjava/
-java -cp $(for i in plibjava/*.jar ; do echo -n $i\; ; done).  com.fuelbigdata.publisher.ProducerApp
-rm -rf plibjava
+if [ "$(uname)" == "MINGW64_NT-10.0-19042" ]; then
+    mkdir -p plibjava
+    cp -rf  ~/.m2/repository/javax/jms/jms/1.1/jms-1.1.jar plibjava/
+    cp -rf kafka-jms-fatjar/target/kafka-jms-client-fat-4.1.2.jar plibjava/
+    cp -rf target/jmstokafka-0.0.1.jar plibjava/
+    java -cp $(for i in plibjava/*.jar ; do echo -n $i\; ; done).  com.fuelbigdata.publisher.ProducerApp
+    rm -rf plibjava
+else
+    java -classpath ~/.m2/repository/javax/jms/jms/1.1/jms-1.1.jar:kafka-jms-fatjar/target/kafka-jms-client-fat-4.1.2.jar:target/jmstokafka-0.0.1.jar com.fuelbigdata.publisher.ProducerApp
+fi
 #
